examples/ai-chatbot/requirements.txt
# Anthropic API
anthropic==0.39.0

# CLI formatting
rich==13.7.0

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-cov==4.1.0
